client_app_sample/pyproject.toml
[project]
name = "client-app-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "pydantic-settings>=2.9.1",
    "redis>=6.0.0",
]
